✨ feat: Add nextcloud-with-smbclient image (#24)
* ✨ feat: Add nextcloud-with-smbclient image

This commit adds a new Docker image for Nextcloud with the SMB client
extension enabled. The image is based on the official Nextcloud image and
installs the necessary dependencies to enable the SMB client functionality.

The changes include:

- Added a README.md file with instructions for building and running the
  Docker image.
- Created a GitHub Actions workflow to automatically build and push the
  Docker image to DockerHub.
- Added a Dockerfile to build the Docker image, which installs the SMB
  client extension and enables it in the PHP configuration.
- Added a VERSION file with the current version of the image.

The purpose of these changes is to provide a ready-to-use Nextcloud
image that includes the SMB client functionality, making it easier for
users to integrate Nextcloud with their SMB file shares.

* ✨🐳 feat(nextcloud): Add smbclient support to Nextcloud

Adds support for smbclient to the Nextcloud Docker image. This allows
Nextcloud to access SMB/CIFS shares directly without the need for
external plugins or third-party integrations. The changes include:

- Update the base image to the latest 30.0-apache version
- Install the necessary build dependencies for the smbclient PHP
  extension
- Install and enable the smbclient PHP extension
- Clean up the APT cache and temporary files to reduce the image size