Reorganize Docker Apps and Implement README Generation (#41)
* 🎨 feat: add readme-generator script and config files

This commit introduces a new script to automatically generate the README.md file
based on the available apps in the "Apps" directory. The script fetches the
app configurations from the GitHub repository and generates a markdown table
with the relevant information.

Additionally, this commit adds configuration files for the following apps:

- BTOP
- Genmon
- Nextcloud with SMBClient

The configuration files contain metadata about each app, such as the ID, name,
Docker image, version, YouTube video, and documentation link. This information
is used by the readme-generator script to populate the README.md file.

The commit also includes a new GitHub Actions workflow that runs the
readme-generator script on a daily schedule and creates a pull request with the
updated README.md file. The pull request is automatically labeled as "automerge"
to streamline the update process.

* ✨ feat: Add project dependencies and configuration

This commit adds the necessary dependencies and configuration files for the
big-bear-casaos project. The key changes include:

- Adding `package.json` with project metadata, scripts, and dev dependencies
- Adding `pnpm-lock.yaml` to manage package dependencies with pnpm
- Configuring linting, testing, and commit message conventions

These changes set up the project's development environment and ensure
consistent code quality and collaboration across the team.

* 🔧 Move docker apps to Apps directory

This is implemented for the README generation.

* ✨ feat: update repository and project details

Updates the repository URL, author, license, and project description in the
package.json file to reflect the correct project details.

* 🔖 docs(README): enhance project description

This commit enhances the project description in the README.md file to
provide more detailed and informative information about the Docker images
maintained by BigBearTechWorld. The changes focus on highlighting the
careful design, optimization, and maintenance of the images, making them
suitable for both development and production environments.

* 🔧 refactor: Restructure Docker build context and Dockerfile paths

This commit restructures the Docker build context and Dockerfile paths for several applications in the repository. The changes are as follows:

- The build context for `btop`, `nextcloud-with-smbclient`, and `pocketbase` has been changed from the root directory to the `Apps` subdirectory.
- The Dockerfile paths for the above applications have been updated to reflect the new build context.
- The paths to the `VERSION` files for various applications have been updated to the new `Apps` subdirectory.

These changes are made to better organize the repository structure and improve the maintainability of the Docker build process.