✨feat: Add quick start section and featured apps list to README
This commit adds a new "Quick Start" section to the README, providing instructions
for pulling the Docker images. It also introduces a "Featured Apps" section,
which will display a list of the available Docker images. These changes aim to
improve the user experience and make it easier for users to discover and
utilize the provided Docker images.