🐻‍❄️ feat(btop): Add authentication support and update dependencies (#67)
* 🐻‍❄️ feat(btop): Add authentication support and update dependencies

This commit adds support for authentication to the btop container, allowing users to
set a username and password for accessing the web-based terminal. Additionally, the
Dockerfile and dependencies have been updated to use the latest versions.

The key changes include:

- Added environment variables for configuring authentication (user, password, and
  enabled status) in both the Dockerfile and the entrypoint script.
- Implemented logic in the entrypoint script to handle authentication, either
  enabling it with the provided credentials or running the container without
  authentication.
- Updated the base image for the Dockerfile from `alpine:latest` to `alpine:3.21.3`.
- Bumped the version of the btop application from `0.0.8` to `0.0.9`.

These changes improve the security and maintainability of the btop container, making
it easier to deploy and manage in production environments.

* 🔒 feat: disable authentication for TTYD and GOTTY

Disables the authentication for the TTYD and GOTTY containers by setting
the `TTYD_AUTH_ENABLED` and `GOTTY_AUTH_ENABLED` environment variables to
`false`. This change is made to simplify the deployment and usage of these
containers, as authentication may not be required in all scenarios.

* Add CHANGELOG.md for btop

Added a changelog and tried to add former changes.

* 🔒 feat(auth): disable authentication by default

Disable authentication by default for the btop application. This change
is made to improve the user experience and make the application more
accessible to users who do not require authentication. The environment
variables `GOTTY_AUTH_ENABLED` and `TTYD_AUTH_ENABLED` are now set to
"false" by default.

* ✨ feat: Remove CasaOS integration for btop monitoring tool on CHANGELOG.md

Edit the CHANGELOG.md and remove the CasaOS Reference.

* ✨ feat(Apps/btop): Implement secure authentication for btop web interface

Adds basic authentication to protect the btop web interface. Implements proper
credential handling in the entrypoint scripts by using quoted arguments to
prevent word splitting. Also uses ARG instructions for sensitive data with ENV
overrides to improve security. Fixes a potential command injection
vulnerability in the entrypoint scripts.